requirement.txt
camel-ai
streamlit